backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Session==0.5.0
python-dotenv==1.0.0
cachelib==0.13.0
# psycopg2-binary==2.9.9  # Uncomment when using PostgreSQL
pytest==8.4.2
pytest-flask==1.3.0
